Extract featurization code from DocumentSegmenter to Featurizer
This makes both halves easier to test. It's now possible to test the
tokenization directly, instead of trying to infer from the feature
vectors.  It's also possible to feed token sets directly to the
featurizer rather indirectly through the segmentation/tokenization code.
